Add files via upload
The AES project aims to develop and implement a ML model capable of analyzing and evaluating essays. A generic model for AES using GloVe embeddings and
traditional models (Linear regression and SVM) is cretaed for evaluating the output. The GloVe embeddings glove.6b.50d are used for the AES using GloVe. The other GloVe embeddings with different dimensions (100 or 300) can be used as well.